requirements.txt
channels==4.1.0
Django==5.0.4
Django==5.1
djangorestframework==3.15.1
PyJWT==2.3.0
PyJWT==2.9.0
pytest==8.2.2
python-decouple==3.8
pytz==2024.1
pytz==2023.4
pytz==2022.1
Requests==2.32.3
